# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: BBS Demo REST API Specification
  description: Spring REST Docs 기반 OpenAPI 3.0 자동 추출 명세서
  version: 1.0.0
servers:
- url: http://localhost:8080
tags: []
paths:
  /api/boards:
    get:
      tags:
      - 1. 게시판 API
      summary: 게시판 전체 목록 조회
      description: 등록된 모든 게시판 목록을 조회합니다.
      operationId: board-get-all
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-boards-33084672"
              examples:
                board-get-all:
                  value: "[{\"id\":1,\"name\":\"공지사항\",\"description\":\"공지사항 게시판\"\
                    ,\"createdAt\":\"2026-08-22T21:00:00\",\"updatedAt\":\"2026-08-22T21:00:00\"\
                    }]"
    post:
      tags:
      - 1. 게시판 API
      summary: 게시판 생성
      description: 새로운 게시판을 생성합니다.
      operationId: board-create
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/api-boards-33084672"
            examples:
              board-create:
                value: "{\"name\":\"자유게시판\",\"description\":\"자유 게시판입니다.\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-boards-33084672"
              examples:
                board-create:
                  value: "{\"id\":1,\"name\":\"자유게시판\",\"description\":\"자유 게시판입니다\
                    .\",\"createdAt\":\"2026-08-22T21:00:00\",\"updatedAt\":\"2026-08-22T21:00:00\"\
                    }"
  /api/boards/{boardId}:
    delete:
      tags:
      - 1. 게시판 API
      summary: 게시판 삭제
      description: 게시판 및 하위 게시글을 삭제합니다.
      operationId: board-delete
      parameters:
      - name: boardId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "204":
          description: "204"
  /api/posts/{postId}:
    get:
      tags:
      - 2. 게시글 API
      summary: 게시글 상세 조회
      description: 게시글 상세 내용과 댓글 목록을 함께 조회합니다.
      operationId: post-get-detail
      parameters:
      - name: postId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-boards-33084672"
              examples:
                post-get-detail:
                  value: "{\"id\":10,\"boardId\":1,\"title\":\"조회 제목\",\"content\"\
                    :\"조회 내용\",\"writer\":\"홍길동\",\"comments\":[{\"id\":100,\"content\"\
                    :\"좋은 글입니다!\",\"writer\":\"이순신\",\"createdAt\":\"2026-08-22T21:15:00\"\
                    ,\"updatedAt\":\"2026-08-22T21:15:00\"}],\"createdAt\":\"2026-08-22T21:10:00\"\
                    ,\"updatedAt\":\"2026-08-22T21:10:00\"}"
  /api/boards/{boardId}/posts:
    post:
      tags:
      - 2. 게시글 API
      summary: 게시글 작성
      description: 특정 게시판에 게시글을 작성합니다.
      operationId: post-create
      parameters:
      - name: boardId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/api-boards-33084672"
            examples:
              post-create:
                value: "{\"title\":\"제목\",\"content\":\"내용\",\"writer\":\"작성자\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-boards-33084672"
              examples:
                post-create:
                  value: "{\"id\":10,\"boardId\":1,\"title\":\"제목\",\"content\":\"\
                    내용\",\"writer\":\"작성자\",\"comments\":[],\"createdAt\":\"2026-08-22T21:10:00\"\
                    ,\"updatedAt\":\"2026-08-22T21:10:00\"}"
  /api/posts/{postId}/comments:
    post:
      tags:
      - 3. 댓글 API
      summary: 댓글 작성
      description: 특정 게시글에 댓글을 작성합니다.
      operationId: comment-create
      parameters:
      - name: postId
        in: path
        description: ""
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: "#/components/schemas/api-boards-33084672"
            examples:
              comment-create:
                value: "{\"content\":\"좋은 글입니다!\",\"writer\":\"이순신\"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/api-boards-33084672"
              examples:
                comment-create:
                  value: "{\"id\":100,\"content\":\"좋은 글입니다!\",\"writer\":\"이순신\"\
                    ,\"createdAt\":\"2026-08-22T21:15:00\",\"updatedAt\":\"2026-08-22T21:15:00\"\
                    }"
components:
  schemas:
    api-boards-33084672:
      type: object